requirements.txt
# Core dependencies
click==8.1.7
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
onnxruntime==1.16.0
Pillow==10.1.0
opencv-python==4.8.1
ffmpeg-python==0.2.0
scikit-learn==1.3.0

# File handling
python-magic==0.4.27
chardet==5.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
colorama==0.4.6
tqdm==4.66.1

# Logging
structlog==23.2.0